+++ email2slack/__main__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+
+import email2slack
+
+if __name__ == '__main__':
+    email2slack.main()
